scripts/requirements.txt
pytorch
numpy
scipy
bandmat
